[Feat] BattleManagerState BattleStart, TurnSequence
배틀매니저의 배틀스타트 상태, 턴 순서 결정 상태 로직 구현

+++ Assets/0.Script/Unit/UnitDataSO.cs
@@ -10,7 +10,7 @@ public class UnitDataSO : ScriptableObject
     [field: SerializeField] public int Hp { get; private set; }
     [field: SerializeField] public float AttLevel { get; private set; }
     [field: SerializeField] public float DefLevel { get; private set; }
-    [field: SerializeField] public float Speed { get; private set; }
+    [field: SerializeField] public int Speed { get; private set; }
     
     [SerializeField] private UnitView _unitPrefab;
